Rename hooks and functions to be more consistent
