+++ src/hooks/useTokenSupply.ts
@@ -1,7 +1,7 @@
 import { useTokenSuppliesQuery } from "src/generated/graphql";
 import { getOhmCirculatingSupply, getOhmFloatingSupply } from "src/helpers/subgraph/TreasuryQueryHelper";
 import { getSubgraphUrl } from "src/helpers/SubgraphUrlHelper";
-import { useTokenRecordsLatestBlock } from "src/hooks/useTokenRecords";
+import { useTokenRecordsLatestBlock } from "src/hooks/useTokenRecordsMetrics";
 import { DEFAULT_RECORD_COUNT } from "src/views/TreasuryDashboard/components/Graph/Constants";
 
 const QUERY_OPTIONS = { refetchInterval: 60000 }; // Refresh every 60 seconds
