+++ src/views/TreasuryDashboard/components/Metric/Metric.tsx
@@ -10,7 +10,7 @@ import {
   useTotalValueDeposited,
 } from "src/hooks/useProtocolMetrics";
 import { useStakingRebaseRate } from "src/hooks/useStakingRebaseRate";
-import { useTreasuryMarketValue } from "src/hooks/useTokenRecords";
+import { useTreasuryMarketValue } from "src/hooks/useTokenRecordsMetrics";
 import { useOhmCirculatingSupply, useOhmFloatingSupply } from "src/hooks/useTokenSupply";
 import { useLiquidBackingPerGOhm, useLiquidBackingPerOhmFloating, useMarketCap } from "src/hooks/useTreasuryMetrics";
